Updating files and adding the 2nd part of the analysis
I decided to divide this analysis into 2 parts, which is why I updated the name of the initial file.

The first part is to analyze the data by "Type", and the second part is to analyze the data by "Tags".

The database had presented a problem with the name of the Tag column, there was a space in the name, so I corrected this problem and updated the document.